refact(product): AdminCategory DTO 통합
